5. First endpoints stats working

+++ src/main/java/com/example/restaurantapp/service/ReservationService.java
@@ -1,5 +1,6 @@
 package com.example.restaurantapp.service;
 
+import com.example.restaurantapp.dto.CustomerReservationDto;
 import com.example.restaurantapp.model.Reservation;
 import com.example.restaurantapp.repository.ReservationRepository;
 import lombok.RequiredArgsConstructor;
@@ -15,39 +16,46 @@ public class ReservationService {
     private final ReservationRepository reservationRepository;
 
 
-    /**
-     *
-     * @return
-     */
-    public List<Reservation> findAllReservations() {
-        return reservationRepository.findAll();
+    public CustomerReservationDto createReservation(CustomerReservationDto dto) {
+        Reservation reservation = this.convertToEntity(dto);
+        reservation = reservationRepository.save(reservation);
+        return converToDto(reservation);
     }
 
-    /**
-     *
-     * @param id
-     * @return
-     */
-    public Optional<Reservation> findReservationById(Long id) {
-        return reservationRepository.findById(id);
-    }
-
-    /**
-     *
-     * @param reservation
-     * @return
-     */
-    public Reservation createReservation(Reservation reservation) {
-        return reservationRepository.save(reservation);
-    }
-
-    /**
-     *
-     * @param id
-     */
-    public void deleteReservation(Long id) {
-        reservationRepository.deleteById(id);
+    private CustomerReservationDto converToDto(Reservation reservation) {
+        CustomerReservationDto customerReservationDto = new CustomerReservationDto();
+        customerReservationDto.setReservationTime(reservation.getReservationTime());
+        customerReservationDto.setCustomerName(reservation.getCustomerName());
+        customerReservationDto.setId(reservation.getId());
+        customerReservationDto.setCustomerEmail(reservation.getCustomerEmail());
+        return customerReservationDto;
     }
 
 
+    private Reservation convertToEntity(CustomerReservationDto dto){
+        Reservation reservation = new Reservation();
+        reservation.setReservationTime(dto.getReservationTime());
+        reservation.setCustomerEmail(dto.getCustomerEmail());
+        reservation.setCustomerName(dto.getCustomerName());
+        return reservation;
+    }
+
+    public Boolean deleteReservation(Long id) {
+        Optional<Reservation> reservation = reservationRepository.findById(id);
+        if (reservation.isPresent()){
+            reservationRepository.deleteById(id);
+            return true;
+        } else {
+            return false;
+        }
+    }
+
+    public CustomerReservationDto getById(Long id) {
+        Optional<Reservation> reservation = reservationRepository.findById(id);
+        CustomerReservationDto customerReservationDto = null;
+        if(reservation.isPresent()){
+            customerReservationDto = converToDto(reservation.get());
+        }
+        return customerReservationDto;
+    }
 }
